\name{correctCdt}
\alias{correctCdt}
\title{ Function corrects the CDT file generated by posthoc.}
\description{
	This function is used to correct the .cdt file generated
  by the "posthoc" executable.
}
\usage{
correctCdt(geneDataFromTxt, fileName="", nreplicates=1, atr=NULL)
}
\arguments{
\item{geneDataFromTxt}{
		This object represents the original gene data.
	}
\item{fileName}{
		This file should be present in the current
		folder. The filename can contain anything, but
		only the name of teh file without any path and
		extension will be used to track .cdt file.
	}
\item{nreplicates}{
	}
\item{atr}{  
	   If NULL, this paramter is ignored. Otherwise, `atr` is assumed
	   to be a hierarchical clustering (hclust object) of the samples (columns)
	   represented in the .cdt file.
	}
}
\details{
	
}
\value{
  This functions returns a matrix object containing
	all the rows in .cdt file.
}
\author{ Vinayak Kumar, Mario Medvedovic}
\note{ Email Address : kumarvk@ucmail.uc.edu }
\seealso{processPostHocParas, runGimmNPosthoc, processGimmParas}
\keyword{ cluster }

