\name{GalData}
\alias{GalData}
\alias{ExpressionChip}
\title{
  Contains table object data which can be used in runGimmNPosthoc().
}
\description{
  This object contains a table object which can be used in
	either runGimmNPosthoc() or runGimm().
}
\section{References}{
}
\author{ Vinayak Kumar, Mario Medvedovic}
\note{ Email Address : kumarvk@ucmail.uc.edu }
\keyword{cluster}

