\name{getOnlyFileName}
\alias{getOnlyFileName}
\title{ Funciton to extract filename without any extension and path.}
\description{
This function retrives the filename from the string provided as an argument wituhout any extension and path information 
}
\usage{
getOnlyFileName(filePath)
}
\arguments{
  \item{filePath}{This is the path and the filename enterd by the user }
}
\details{
}
\value{
  It returns the filename without any extension
}
\author{ Vinayak Kumar, Mario Medvedovic}
\note{ Email Address : kumarvk@ucmail.uc.edu }
\seealso{correctCdt, convTxtToInp, processPostHocParas, runGimmNPosthoc, processGimmParas}
\keyword{ cluster }

