\name{getPerm}
\alias{getPerm}
\title{Gets all the permutations from .gtr file}
\description{
 Gets all the permutations from .gtr file
}
\usage{
	getPerm(merge, clusterSize, level, start, perm)
}
\arguments{
  \item{merge}{ }
  \item{clusterSize}{ }
  \item{level}{ }
  \item{start}{ }
  \item{perm}{ }
}
\details{
}
\value{
}
\author{Vinayak Kumar, Mario Medvedovic}
\note{Email Address : kumarvk@ucmail.uc.edu }
\seealso{importGtr}
\keyword{ cluster }

