\name{gimmRFromPos}
\alias{gimmRFromPos}
\title{
	Method to create gimmR object corresponding to .pos file.
}
\description{
	This method first creates the table object from .pos file,
	and then uses this object as input to runGimmNPosthoc().
	Before running this method, user must run either runGimm()
	and runPosthoc() or runGimmNPosthoc(), so that .pos file
	is generated.
}
\usage{
gimmRFromPos(posDataFileName, M, T, nIter=10000, nreplicates=1,
				 contextSpecific="n", nContexts=1, contextLengths=NA,
				 clusterShape="v", burnIn=5000, elipticalWithin="n", matrixOut="n",
				 readingRaw="y", probsOut="y", clusterNumber=2,
				 clusterOption="average", verbose=FALSE, intFiles=FALSE)
}
\arguments{
  \item{posDataFileName}{
		Name of the .pos file. if not provided, default fileName
		is taken as "Result.pos". If file not found, method throws
		an error.
	}
  \item{M}{
		Dimensionality of the expression vectors to be clustered. I.e. the number
		of microarray. Same as given to runGimmNPosthoc() before running
		this method.
	}
  \item{T}{
		Number of genes to be clustered. Same as given to runGimmNPosthoc()
		before running this method.
	}
  \item{nIter}{ 
		Number of iterations to be generated by the Gibbs sampler. Same as
		given to runGimmNPosthoc() before running	this method.
	}
  \item{nreplicates}{
		Number of experimental replicates. Same as
		given to runGimmNPosthoc() before running	this method.
  }
  \item{contextSpecific}{
		This parameter specifies whether the context-specific model to
		be used. If set to "y", the nContexts and contextLengths need
		also to be specified.
	}
  \item{nContexts}{
		Number of contexts to be used in the context-specific clustering.
		If nContexts>1 is specified, the context-specific model is assumed.
		  }
  \item{contextLengths}{
		Vector specifying lengths of the contexts in the context-specific
		clustering. It has to satisfy following equality
		sum(contextLenghts)==M
		  }
  \item{clusterShape}{  
		"v" requests the model with different variances for different clusters.
		"e" assumes equal variances for all clusters.
		}
  \item{burnIn}{
		Number of Gibbs sampler iterations to be discarded as "burn-in".
		}
  \item{elipticalWithin}{
		This parameter indicates the structure of covariance matrix
		in model. The covariance matrix defines the correlation of
		observations from different experimental conditions. For
		experimental design model, its value is "d", for compound 
		symmetry model, its value is "c" while for unstructured model;
		its value is "n".
	}
  \item{matrixOut}{  
		Not used
		}
  \item{readingRaw}{  
		Not used
		}
  \item{probsOut}{
		Not used
		}
  \item{clusterNumber}{
		Not used
		}
  \item{clusterOption}{
		Linkage principle to be used in constructing hierarchical clustering.
		"average" and "complete" are supported.
		 }
  \item{verbose}{
		If true, all the internal cmments of the executables
		will be displayed on console.
	}
  \item{intFiles}{
		If true, the internal files generated by the executables
		gimm and posthoc won't be deleted.
  }
}
\details{

}

\author{ Vinayak Kumar, Mario Medvedovic }
\note{ For additional information visit http://eh3.uc.edu/gimm }
\seealso{  }
\keyword{ cluster }
