\name{readMpm}
\alias{readMpm}
\title{ Generates the Table object for .mpm file.}
\description{
	This function is used to generate the table object 
	from .mpm file.
}
\usage{
readMpm(dataFile)
}
\arguments{
  \item{dataFile}{
		This file should be present in the current
		folder. If this parameter is missing, then
		dataFile is taken as "./Result.mpm"
	}
}
\details{
		The table object generated by this method contains data
		from .mpm file which will be generated when either
		runGimmNPosthoc() or runGimm()and runPosthoc() are
	  executed. So before running this method either of them
		should be executed. 
}
\value{
  This functions returns a table object.
}
\author{ Vinayak Kumar, Mario Medvedovic}
\note{ Email Address : kumarvk@ucmail.uc.edu }
\seealso{processPostHocParas, runGimmNPosthoc, processGimmParas}
\keyword{ cluster }


